+++ packages/kb-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # kb-server
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @kb/core@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes